+++ system/include/emscripten/trace.h
@@ -9,6 +9,7 @@
 
 #include <stdbool.h>
 #include <stdint.h>
+#include <emscripten/em_macros.h>
 
 #ifdef __cplusplus
 extern "C" {
@@ -16,55 +17,55 @@ extern "C" {
 
 #ifdef __EMSCRIPTEN_TRACING__
 
-void emscripten_trace_configure(const char *collector_url, const char *application);
+void EM_IMPORT_WRAP(emscripten_trace_configure)(const char *collector_url, const char *application);
 
-void emscripten_trace_configure_for_google_wtf(void);
+void EM_IMPORT_WRAP(emscripten_trace_configure_for_google_wtf)(void);
 
-void emscripten_trace_configure_for_test(void);
+void EM_IMPORT_WRAP(emscripten_trace_configure_for_test)(void);
 
-void emscripten_trace_set_enabled(bool enabled);
+void EM_IMPORT_WRAP(emscripten_trace_set_enabled)(bool enabled);
 
-void emscripten_trace_set_session_username(const char *username);
+void EM_IMPORT_WRAP(emscripten_trace_set_session_username)(const char *username);
 
-void emscripten_trace_record_frame_start(void);
+void EM_IMPORT_WRAP(emscripten_trace_record_frame_start)(void);
 
-void emscripten_trace_record_frame_end(void);
+void EM_IMPORT_WRAP(emscripten_trace_record_frame_end)(void);
 
-void emscripten_trace_mark(const char *message);
+void EM_IMPORT_WRAP(emscripten_trace_mark)(const char *message);
 
-void emscripten_trace_log_message(const char *channel, const char *message);
+void EM_IMPORT_WRAP(emscripten_trace_log_message)(const char *channel, const char *message);
 
-void emscripten_trace_report_error(const char *error);
+void EM_IMPORT_WRAP(emscripten_trace_report_error)(const char *error);
 
-void emscripten_trace_record_allocation(const void *address, int32_t size);
+void EM_IMPORT_WRAP(emscripten_trace_record_allocation)(const void *address, int32_t size);
 
-void emscripten_trace_record_reallocation(const void *old_address, const void *new_address, int32_t size);
+void EM_IMPORT_WRAP(emscripten_trace_record_reallocation)(const void *old_address, const void *new_address, int32_t size);
 
-void emscripten_trace_record_free(const void *address);
+void EM_IMPORT_WRAP(emscripten_trace_record_free)(const void *address);
 
-void emscripten_trace_annotate_address_type(const void *address, const char *type);
+void EM_IMPORT_WRAP(emscripten_trace_annotate_address_type)(const void *address, const char *type);
 
-void emscripten_trace_associate_storage_size(const void *address, int32_t size);
+void EM_IMPORT_WRAP(emscripten_trace_associate_storage_size)(const void *address, int32_t size);
 
-void emscripten_trace_report_memory_layout(void);
+void EM_IMPORT_WRAP(emscripten_trace_report_memory_layout)(void);
 
-void emscripten_trace_report_off_heap_data(void);
+void EM_IMPORT_WRAP(emscripten_trace_report_off_heap_data)(void);
 
-void emscripten_trace_enter_context(const char *name);
+void EM_IMPORT_WRAP(emscripten_trace_enter_context)(const char *name);
 
-void emscripten_trace_exit_context(void);
+void EM_IMPORT_WRAP(emscripten_trace_exit_context)(void);
 
-void emscripten_trace_task_start(int task_id, const char *name);
+void EM_IMPORT_WRAP(emscripten_trace_task_start)(int task_id, const char *name);
 
-void emscripten_trace_task_associate_data(const char *key, const char *value);
+void EM_IMPORT_WRAP(emscripten_trace_task_associate_data)(const char *key, const char *value);
 
-void emscripten_trace_task_suspend(const char *explanation);
+void EM_IMPORT_WRAP(emscripten_trace_task_suspend)(const char *explanation);
 
-void emscripten_trace_task_resume(int task_id, const char *explanation);
+void EM_IMPORT_WRAP(emscripten_trace_task_resume)(int task_id, const char *explanation);
 
-void emscripten_trace_task_end(void);
+void EM_IMPORT_WRAP(emscripten_trace_task_end)(void);
 
-void emscripten_trace_close(void);
+void EM_IMPORT_WRAP(emscripten_trace_close)(void);
 
 #else