+++ system/lib/compiler-rt/lib/sanitizer_common/sanitizer_linux.cpp
@@ -1134,10 +1134,6 @@ uptr GetPageSize() {
 }
 #endif // !SANITIZER_ANDROID
 
-#if SANITIZER_EMSCRIPTEN
-extern "C" uptr emscripten_get_module_name(char *buf, uptr buf_len);
-#endif
-
 #if !SANITIZER_OPENBSD
 uptr ReadBinaryName(/*out*/char *buf, uptr buf_len) {
 #if SANITIZER_SOLARIS