+++ system/lib/compiler-rt/lib/sanitizer_common/sanitizer_symbolizer_emscripten.cpp
@@ -16,16 +16,11 @@
 
 #if SANITIZER_EMSCRIPTEN
 
+#include <emscripten.h>
+
 #include "sanitizer_symbolizer_internal.h"
 
 namespace __sanitizer {
-
-extern "C" {
-  const char *emscripten_pc_get_function(uptr pc);
-  const char *emscripten_pc_get_file(uptr pc);
-  int emscripten_pc_get_line(uptr pc);
-  int emscripten_pc_get_column(uptr pc);
-}
 
 class EmscriptenSymbolizerTool : public SymbolizerTool {
  public: